fix #1129 empty only existing tables in emptyAll()

+++ lib/Db/CoreRequestBuilder.php
@@ -1164,10 +1164,11 @@ class CoreRequestBuilder {
 	 */
 	public function emptyAll() {
 		foreach ($this->tables as $table) {
-			$qb = $this->dbConnection->getQueryBuilder();
-			$qb->delete($table);
-
-			$qb->execute();
+			if ($schema->hasTable($table)) {
+				$qb = $this->dbConnection->getQueryBuilder();
+				$qb->delete($table);
+				$qb->execute();
+			}
 		}
 	}
